Some comments so pull request can be created.
Forgot to create the empty repository at the beginning of the project. Now need to create a Pull Request, so changes have to be made. Realize now that I should probabaly get in a habbit of branching when I start working on something and then merging into main when everything works. Live and learn.